Fix collision with _make_attachment() method in 4.4
This changes the name of our helper function, `_make_attachment()`
to `create_upload_object` to avoid a name collision with a public
method introduced to the `WP_UnitTestCase` class during the 4.4
cycle.

I also mimicked the test performance improvements in core that uses
a single upload for all the tests, where applicable so that we don't
have the overhead of uploading an image during each test.

Finally, this changes the main class name for our tests to
`RICG_Responsive_Images_Tests`.