Update srcset tests to check for 'medium_large'
The 'medium_large' size was added to core during the 4.4 cycle,
which caused our expected `srcset` attributes to be off. This adds
a whitelist for testing the default soft crops if they exist, which
will make use of the 'meduim_large' size when applicable and still
work when testing against versions of WP before the image size was
added.