Improved calculation of image ratios for srcset attributes
Uses wp_constrain_dimensions() to calculate the expected dimensions
of the larger size being compared if it were resized to the width of
the smaller image. By comparing the expected dimensions to the actual
dimensions of the smaller image, we should be able to reliably tell if
the current image size has the same aspect ratio and, if so, include it
in the srcset attribute.

See: https://core.trac.wordpress.org/ticket/34955